@@ -4,7 +4,7 @@
   "type": "library",
   "license": "GPL-3.0-or-later",
   "require": {
-    "php": "7.4.* | 8.0.* | 8.1.* ",
+    "php": "8.1.*",
     "nette/utils": "^v3.2.0",
     "nette/di": "^3.0.0",
     "nette/application": "^3.1.0",
@@ -16,7 +16,7 @@
     "ext-gettext": "*"
   },
   "require-dev": {
-    "squizlabs/php_codesniffer": "3.6.0",
+    "squizlabs/php_codesniffer": "3.7.1",
     "nette/tester": "^v2.4.0"
   },
   "authors": [
