@@ -10,25 +10,28 @@ class Message
 {
     use SmartObject;
 
+    /** @deprecated */
     public const LVL_ERROR = 'danger';
+    /** @deprecated */
     public const LVL_WARNING = 'warning';
+    /** @deprecated */
     public const LVL_SUCCESS = 'success';
+    /** @deprecated */
     public const LVL_INFO = 'info';
+    /** @deprecated */
     public const LVL_PRIMARY = 'primary';
 
-    public string $text;
-    public string $level;
+    public readonly MessageLevel $level;
 
-    public function __construct(string $message, string $level)
+    public function __construct(public readonly string $message, MessageLevel|string $level)
     {
-        $this->text = $message;
-        $this->level = $level;
+        $this->level = ($level instanceof MessageLevel) ? $level : MessageLevel::from($level);
     }
 
     public function __toArray(): array
     {
         return [
-            'text' => $this->text,
+            'text' => $this->message,
             'level' => $this->level,
         ];
     }
